Rename bpf/lib/aux.h to auxvars.h
This filename makes the Cilium repository impossible to check out
on Windows because 'aux' is one of legacy special filenames.
