+++ bpf/lib/auxvars.h
